Delete old tests
